fix fix double error dialog when replacing text by text selection and typing
